Renamed and Restructured Notes to Account Notes
