+++ src/main/java/com/salessparrow/api/lib/crmActions/createAccountNote/CreateNoteInterface.java
@@ -1,4 +1,4 @@
-package com.salessparrow.api.lib.crmActions.createNote;
+package com.salessparrow.api.lib.crmActions.createAccountNote;
 
 import org.springframework.stereotype.Component;
