+++ src/main/java/com/salessparrow/api/lib/crmActions/getAccountNoteDetails/GetAccountNoteDetails.java
@@ -1,4 +1,4 @@
-package com.salessparrow.api.lib.crmActions.getNoteDetails;
+package com.salessparrow.api.lib.crmActions.getAccountNoteDetails;
 
 import org.springframework.stereotype.Component;
 
@@ -9,6 +9,6 @@ import com.salessparrow.api.dto.formatter.GetNoteDetailsFormatterDto;
  * GetNoteDetails is an interface for the GetNoteDetails action for the CRM.
  */
 @Component
-public interface GetNoteDetails {
+public interface GetAccountNoteDetails {
     public GetNoteDetailsFormatterDto getNoteDetails(User user, String noteId);
 }
