+++ src/main/java/com/salessparrow/api/lib/crmActions/getAccountNotesList/GetAccountNotesList.java
@@ -1,4 +1,4 @@
-package com.salessparrow.api.lib.crmActions.getNotesList;
+package com.salessparrow.api.lib.crmActions.getAccountNotesList;
 
 import org.springframework.stereotype.Component;
 
@@ -9,7 +9,7 @@ import com.salessparrow.api.dto.formatter.GetNotesListFormatterDto;
  * GetNotesList is an interface for the GetNotesList action for the CRM.
  */
 @Component
-public interface GetNotesList {
+public interface GetAccountNotesList {
     public GetNotesListFormatterDto getNotesList(User user, String accountId);
 }
