+++ src/main/java/com/salessparrow/api/services/accountNotes/CreateAccountNoteService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import com.salessparrow.api.domain.SalesforceUser;
 import com.salessparrow.api.dto.formatter.CreateNoteFormatterDto;
 import com.salessparrow.api.dto.requestMapper.NoteDto;
-import com.salessparrow.api.lib.crmActions.createNote.CreateNoteFactory;
+import com.salessparrow.api.lib.crmActions.createAccountNote.CreateNoteFactory;
 
 import jakarta.servlet.http.HttpServletRequest;
 
@@ -14,7 +14,7 @@ import jakarta.servlet.http.HttpServletRequest;
  * CreateNoteService is a service class for the create note action for the CRM.
  */
 @Service
-public class CreateNoteService {
+public class CreateAccountNoteService {
   
   @Autowired
   private CreateNoteFactory createNoteFactory;
